docs: update README and add new CLI scripts for protspace-query and protspace-local
