+++ README.md
@@ -207,7 +207,7 @@ Available shapes: circle, circle-open, cross, diamond, diamond-open, square, squ
   - Generates similarity matrix using pymmseqs
   - Fetches UniProt features automatically
 
-2. **Embeddings/Similarity** (for `protspace-local`)
+2. **Local Embeddings/Similarity** (for `protspace-local`)
   - HDF5 (.h5) for embeddings
   - CSV for similarity matrix
 
@@ -221,10 +221,39 @@ Available shapes: circle, circle-open, cross, diamond, diamond-open, square, squ
 
 ### Output
 
-- JSON containing:
-  - Protein features
-  - Projection coordinates
-  - Visualization state (colors, shapes)
+#### protspace-query
+
+- Directory of parquet files:
+  - projections_data.parquet
+  - projections_metadata.parquet
+  - selected_features.parquet
+    - These are selected features specified using `-m` option, if not using `-m` option, it is exactly the all_features.parquet file.
+  - if used `--keep-tmp` flag the files below are also included:
+    - all_features.parquet (fetched from UniProt)
+    - sequences.fasta (fetched from UniProt)
+    - similarity_matrix.csv (generated by PyMMseqs)
+
+- With `--non-binary` flag (legacy version):
+  - selected_features_projections.json (containes selected features and projections data)
+  - if used `--keep-tmp` flag the files below are also included:
+    - all_features.csv
+    - sequences.fasta
+    - similarity_matrix.csv
+
+#### protspace-local
+
+- Directory of parquet files:
+  - projections_data.parquet
+  - projections_metadata.parquet
+  - selected_features.parquet
+    - These are selected features specified using `-m` option, if not using `-m` option, it is exactly the all_features.parquet file.
+  - if used `--keep-tmp` flag the files below are also included:
+    - all_features.parquet (fetched from UniProt)
+
+- With `--non-binary` flag (legacy version):
+  - selected_features_projections.json (containes selected features and projections data)
+  - if used `--keep-tmp` flag the files below are also included:
+    - all_features.csv
 
 ## Citation
